redone tasks one by one

+++ 0x00-hello_world/6-size.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+
 /**
  * main - Entry point
  *
@@ -6,10 +7,11 @@
  */
 int main(void)
 {
-		printf("Size of a char: %d byte(s)\n", sizeof(char));
-		printf("Size of an int: %d byte(s)\n", sizeof(int));
-		printf("Size of a long int: %d byte(s)\n", sizeof(long int));
-		printf("Size of a long long int: %d byte(s)\n", sizeof(long long int));
-		printf("Size of a float: %d byte(s)\n", sizeof(float));
-		return (0);
+	printf("Size of a char: %li byte(s)\n", sizeof(char));
+	printf("Size of an int: %li byte(s)\n", sizeof(int));
+	printf("Size of a long int: %li byte(s)\n", sizeof(long));
+	printf("Size of a long long int: %li byte(s)\n", sizeof(double));
+	printf("Size of a float: %li byte(s)\n", sizeof(float));
+
+	return (0);
 }
